Repository renovation!
The repository now has a function for the viewmodel to request a sharedflow of BrowseResult objects that have a Success and Error state. If the network request is successful the data will be packaged in a BrowseResult.Success object and emitted into the flow. And if the network errors out a BrowseResult.Error will be emitted.