Retry button and copy button
Added a functional retry button to the network error footer.
There is also a toast that now displays when a network error occurs.

Added a copy button to copy the jpl url from a neo. However the string pulled from the textview is coming out corrupted (i think). More testing is needed but the point is it's in there.