New Fragment + Navigation test
Added a new fragment called FeedFragment with a ViewModel and full  data pipeline to go along with it. (Repository, Retrofit service and all the objects needed to deserialize the json) There is also a new database table for the feed near earth objects to be cached in.

Added a UI test for the MainActivity to test the navigation drawers ability to navigate to the FeedFragment.