Bug fixing the Feed Fragment
Fixed a bug where the errorFlow from the NeoRepository was not being collected by the FeedViewModel.

Changed the way FeedViewModel caches date ranges from the UI. It now puts the date range from the UI in a standby variable, then waits to actually cache it in SharedPreferences only if the network request was successful. This is to keep the displayed date range in sync with the actual data in the database.

Added 2 simple unit tests for the FeedViewModel that test if it collects from the feed flow and error flow correctly.